requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
affine=2.3.0=py_0
appnope=0.1.2=py310h2ec42d9_2
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py310he24745e_1
asttokens=2.0.5=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
black=22.1.0=pyhd8ed1ab_0
bleach=4.1.0=pyhd8ed1ab_0
blosc=1.21.0=he49afe7_0
bokeh=2.4.2=py310h2ec42d9_1
boost-cpp=1.74.0=hdbf7018_7
brotli=1.0.9=h0d85af4_6
brotli-bin=1.0.9=h0d85af4_6
brotlipy=0.7.0=py310h1961e1f_1004
bzip2=1.0.8=h0d85af4_4
c-ares=1.18.1=h0d85af4_0
ca-certificates=2022.6.15=h033912b_0
cairo=1.16.0=he01c77b_1009
cartopy=0.20.2=pypi_0
certifi=2022.6.15=py310h2ec42d9_0
cffi=1.15.0=py310hcc37b68_0
cfitsio=4.0.0=hb20e66c_0
cftime=1.6.0=py310h7f5fb2b_1
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.0.3=py310h2ec42d9_1
click-plugins=1.1.1=py_0
cligj=0.7.2=pyhd8ed1ab_1
cloudpickle=2.0.0=pyhd8ed1ab_0
cryptography=36.0.2=py310hd6fa1ae_1
curl=7.81.0=hf45b732_0
cycler=0.11.0=pyhd8ed1ab_0
cython=0.29.27=py310hba3363e_0
cytoolz=0.11.2=py310h1961e1f_2
dask=2022.5.0=pyhd8ed1ab_0
dask-core=2022.5.0=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.6=h811a1a6_3
debugpy=1.5.1=py310hba3363e_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
distributed=2022.5.0=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.2=pyhd8ed1ab_0
expat=2.4.4=he49afe7_0
fiona=1.8.21=pypi_0
flit-core=3.6.0=pyhd8ed1ab_0
focal-stats=0.0.3=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.13.94=h10f422b_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.29.1=py310he24745e_0
freetype=2.10.4=h4cff582_1
freexl=1.0.6=h0d85af4_0
fribidi=1.0.10=hbcb3906_0
fsspec=2022.3.0=pyhd8ed1ab_0
geomappy=0.0.2=dev_0
geopandas=0.10.2=pypi_0
geos=3.10.2=he49afe7_0
geotiff=1.7.0=h0ca5f94_6
gettext=0.19.8.1=hd1a6beb_1008
giflib=5.2.1=hbcb3906_2
hdf4=4.2.15=hefd3b78_3
hdf5=1.12.1=nompi_h2f0ef1a_103
heapdict=1.0.1=py_0
icu=69.1=he49afe7_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.10.1=py310h2ec42d9_0
importlib_metadata=4.10.1=hd8ed1ab_0
importlib_resources=5.4.0=pyhd8ed1ab_0
ipykernel=6.9.0=py310ha188af9_0
ipython=8.0.1=py310h2ec42d9_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.5=pyhd8ed1ab_0
jbig=2.1=h0d85af4_2003
jedi=0.18.1=py310h2ec42d9_0
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h0d85af4_0
json-c=0.15=hcb556a6_0
jsonschema=4.4.0=pyhd8ed1ab_0
jupyter=1.0.0=py310h2ec42d9_7
jupyter_client=7.1.2=pyhd8ed1ab_0
jupyter_console=6.4.0=pyhd8ed1ab_0
jupyter_core=4.9.1=py310h2ec42d9_1
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_widgets=1.0.2=pyhd8ed1ab_0
kealib=1.4.14=ha22a8b1_3
kiwisolver=1.3.2=py310h2fea185_1
krb5=1.19.2=hcfbf3a7_3
lcms2=2.12=h577c468_0
lerc=3.0=he49afe7_0
libblas=3.9.0=13_osx64_openblas
libbrotlicommon=1.0.9=h0d85af4_6
libbrotlidec=1.0.9=h0d85af4_6
libbrotlienc=1.0.9=h0d85af4_6
libcblas=3.9.0=13_osx64_openblas
libclang=13.0.1=default_he082bbe_0
libcurl=7.81.0=hf45b732_0
libcxx=14.0.6=hce7ea42_0
libdap4=3.20.6=h3e144a0_2
libdeflate=1.8=h0d85af4_0
libedit=3.1.20191231=h0678c8f_2
libev=4.33=haf1e3a3_1
libffi=3.4.2=h0d85af4_5
libgdal=3.4.1=h6c40076_2
libgfortran=5.0.0=9_3_0_h6c81a4c_23
libgfortran5=9.3.0=h6c81a4c_23
libglib=2.70.2=hf1fb8c0_2
libiconv=1.16=haf1e3a3_0
libimagequant=2.17.0=h9bde063_1
libkml=1.3.0=h8fd9edb_1014
liblapack=3.9.0=13_osx64_openblas
libllvm13=13.0.1=hd011deb_0
libnetcdf=4.8.1=nompi_h6609ca0_101
libnghttp2=1.46.0=h6f36284_0
libopenblas=0.3.18=openmp_h3351f45_0
libpng=1.6.37=h7cec526_2
libpq=14.1=hea3049e_1
librttopo=1.1.0=hec60dd8_9
libsodium=1.0.18=hbcb3906_1
libspatialite=5.0.1=h88e7940_14
libssh2=1.10.0=h52ee1ee_2
libtiff=4.3.0=hd146c10_2
libwebp=1.2.2=h28dabe5_0
libwebp-base=1.2.2=h0d85af4_1
libxcb=1.13=h0d85af4_1004
libxml2=2.9.12=h7e28ab6_1
libzip=1.8.0=h8b0c345_1
libzlib=1.2.11=h9173be1_1013
llvm-openmp=14.0.4=ha654fa7_0
locket=1.0.0=pyhd8ed1ab_0
lz4=4.0.0=py310h2e8d50f_2
lz4-c=1.9.3=he49afe7_1
markupsafe=2.0.1=py310he24745e_1
matplotlib-base=3.5.1=py310hfc59cd6_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py310he24745e_1005
msgpack-python=1.0.3=py310h6be76da_1
munch=2.5.0=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
mypy=0.961=py310h6c45266_0
mypy_extensions=0.4.3=py310h2ec42d9_4
mysql-common=8.0.28=h694c41f_0
mysql-libs=8.0.28=h115446f_0
nbclient=0.5.10=pyhd8ed1ab_1
nbconvert=6.4.1=py310h2ec42d9_0
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.3=he49afe7_0
nest-asyncio=1.5.4=pyhd8ed1ab_0
netcdf4=1.5.8=nompi_py310h1579a17_101
notebook=6.4.8=pyha770c72_0
nspr=4.32=hcd9eead_1
nss=3.74=h31e2bf1_0
numpy=1.22.2=py310hfbbbacf_0
openjpeg=2.4.0=h6e7aa92_1
openssl=1.1.1p=hfe4f2af_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.0=py310hdd25497_0
pandoc=2.17.1.1=h694c41f_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
partd=1.2.0=pyhd8ed1ab_0
pathspec=0.9.0=pyhd8ed1ab_0
patsy=0.5.2=pyhd8ed1ab_0
pcre=8.45=he49afe7_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.0.1=py310h9351100_0
pip=22.0.3=pyhd8ed1ab_0
pixman=0.40.0=hbcb3906_0
platformdirs=2.4.1=pyhd8ed1ab_1
poppler=21.11.0=h9573804_0
poppler-data=0.4.11=hd8ed1ab_0
postgresql=14.1=he8fe76e_1
proj=8.2.1=h1512c50_0
prometheus_client=0.13.1=pyhd8ed1ab_0
prompt-toolkit=3.0.26=pyha770c72_0
prompt_toolkit=3.0.26=hd8ed1ab_0
psutil=5.9.0=py310h1961e1f_1
pthread-stubs=0.4=hc929b4f_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyproj=3.3.0=py310h4ed7dbf_1
pyqt=5.12.3=py310h2ec42d9_8
pyqt-impl=5.12.3=py310hea9ccf3_8
pyqt5-sip=4.19.18=py310hd046c34_8
pyqtchart=5.12=py310hea9ccf3_8
pyqtwebengine=5.12.1=py310hea9ccf3_8
pyrsistent=0.18.1=py310he24745e_0
pyshp=2.2.0=pypi_0
pysocks=1.7.1=py310h2ec42d9_5
python=3.10.2=h1dd9edd_3_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytz=2021.3=pyhd8ed1ab_0
pyyaml=6.0=py310h1961e1f_4
pyzmq=22.3.0=py310hcce64c8_1
qt=5.12.9=h2a607e2_5
qtconsole=5.2.2=pyhd8ed1ab_1
qtconsole-base=5.2.2=pyhd8ed1ab_1
qtpy=2.0.1=pyhd8ed1ab_0
rasterio=1.2.10=py310h4193d74_4
readline=8.1=h05e3726_0
requests=2.27.1=pyhd8ed1ab_0
rioxarray=0.9.1=pyhd8ed1ab_0
scikit-learn=1.1.1=py310hfc06b38_0
scipy=1.8.0=py310h47774c9_0
seaborn=0.11.2=hd8ed1ab_0
seaborn-base=0.11.2=pyhd8ed1ab_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=60.8.1=py310h2ec42d9_0
shapely=1.8.0=pypi_0
six=1.16.0=pyh6c4a22f_0
sklearn-quantile=0.0.19=dev_0
snuggs=1.4.7=py_0
sortedcontainers=2.4.0=pyhd8ed1ab_0
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.37.0=h23a322b_0
stack_data=0.1.4=pyhd8ed1ab_0
statsmodels=0.13.1=py310h81f86ea_0
tblib=1.7.0=pyhd8ed1ab_0
terminado=0.13.1=py310h2ec42d9_0
testpath=0.5.0=pyhd8ed1ab_0
threadpoolctl=3.1.0=pyh8a188c0_0
tiledb=2.6.2=hca4410a_1
tk=8.6.11=h5dbffcc_1
tomli=2.0.0=pyhd8ed1ab_1
toolz=0.11.2=pyhd8ed1ab_0
tornado=6.1=py310he24745e_2
traitlets=5.1.1=pyhd8ed1ab_0
typed-ast=1.5.2=py310he24745e_0
typing_extensions=4.0.1=pyha770c72_0
tzcode=2021e=h0d85af4_0
tzdata=2021e=he74cb21_0
unicodedata2=14.0.0=py310he24745e_0
urllib3=1.26.9=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.5.2=py310h2ec42d9_1
xarray=2022.3.0=pyhd8ed1ab_0
xerces-c=3.2.3=h6564042_4
xorg-libxau=1.0.9=h35c211d_0
xorg-libxdmcp=1.1.3=h35c211d_0
xz=5.2.5=haf1e3a3_1
yaml=0.2.5=h0d85af4_2
zeromq=4.3.4=he49afe7_1
zict=2.2.0=pyhd8ed1ab_0
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=h9173be1_1013
zstd=1.5.2=h582d3a0_0
